requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
argparse==1.3.0
itsdangerous==0.24
pbr==0.10.7
six==1.9.0
stevedore==1.2.0
wsgiref==0.1.2
